% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-repo.R
\name{repo_names}
\alias{repo_names}
\title{Extract the names of a username}
\usage{
repo_names(username, ...)
}
\arguments{
\item{username}{the username.}

\item{...}{Name-value pairs giving API parameters.}
}
\description{
Extract the names of a username
}
