% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-repo.R
\name{repo_traffic}
\alias{repo_traffic}
\title{Traffic information of a github repository}
\usage{
repo_traffic(owner, repo, ...)
}
\arguments{
\item{owner}{the owner of the repo.}

\item{repo}{the name of the repo.}

\item{...}{Name-value pairs giving API parameters.}
}
\description{
Traffic information of a github repository
}
