% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-repo.R
\name{repo_traffic_overview}
\alias{repo_traffic_overview}
\title{Get an overview with the sum of the traffic data of a username}
\usage{
repo_traffic_overview(username, ...)
}
\arguments{
\item{username}{the username.}

\item{...}{Name-value pairs giving API parameters.}
}
\description{
Get an overview with the sum of the traffic data of a username
}
